requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
alembic==1.9.1
    # via dataset
banal==1.0.6
    # via dataset
click==8.1.3
    # via -r requirements.in
dataset==1.5.2
    # via -r requirements.in
dsnparse==0.1.15
    # via -r requirements.in
greenlet==2.0.1
    # via sqlalchemy
mako==1.2.4
    # via alembic
markupsafe==2.1.1
    # via mako
psycopg2-binary==2.9.5
    # via -r requirements.in
python-dotenv==0.21.0
    # via -r requirements.in
sqlalchemy==1.4.46
    # via
    #   alembic
    #   dataset
